refactor: extract edit cv achievement section to own component
